pyproject.toml
[project]
name = "DeltaDorsal"
version = "0.1.0"
description = "Source code for DeltaDorsal, a 3D hand pose estimation model using dorsal features of the hand"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "lightning>=2.5.3",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "omegaconf>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pip>=25.2",
    "pylogger>=1.1",
    "pyrender>=0.1.45",
    "pyserial>=3.5",
    "pytorch-lightning>=2.5.1.post0",
    "rootutils>=1.0.7",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "setuptools>=80.4.0",
    "smplx>=0.1.28",
    "timm>=1.0.19",
    "torchmetrics>=1.7.1",
    "wandb>=0.21.1",
]

[tool.uv]
package = true
